NTP-376: Test that a TP shows up when `any` tuition type is selected
